+++ cli/src/commands/update/source.rs
@@ -30,9 +30,9 @@ pub struct UpdateSourceArgs {
     /// Set the transform tag of the source
     transform_tag: Option<TransformTag>,
 
-    #[structopt(long = "remove-bucket")]
-    /// Remove bucket from source
-    remove_bucket: bool,
+    #[structopt(long = "detach-bucket")]
+    /// Detach bucket from source
+    detach_bucket: bool,
 }
 
 pub fn update(client: &Client, args: &UpdateSourceArgs, printer: &Printer) -> Result<()> {
@@ -43,7 +43,7 @@ pub fn update(client: &Client, args: &UpdateSourceArgs, printer: &Printer) -> Re
         should_translate,
         bucket,
         transform_tag,
-        remove_bucket,
+        detach_bucket,
     } = args;
 
     let bucket_id = match bucket.to_owned() {
@@ -65,7 +65,7 @@ pub fn update(client: &Client, args: &UpdateSourceArgs, printer: &Printer) -> Re
             .full_name(),
     };
 
-    let bucket_id = if *remove_bucket { None } else { bucket_id };
+    let bucket_id = if *detach_bucket { None } else { bucket_id };
 
     let source = client
         .update_source(